+++ app/src/main/java/com/leeweeder/weighttracker/ui/home/HomeUiState.kt
@@ -4,7 +4,7 @@ import com.leeweeder.weighttracker.domain.model.Log
 
 data class HomeUiState(
     val fiveMostRecentLogs: List<Log> = emptyList(),
-    val goalWeight: Double = 0.0
+    val goalWeight: Float = 0f
 ) {
     val mostRecentLog: Log?
         get() = fiveMostRecentLogs.firstOrNull()
@@ -12,13 +12,13 @@ data class HomeUiState(
     private val previousMostRecentLog: Log?
         get() = if (fiveMostRecentLogs.size > 1) fiveMostRecentLogs[1] else null
 
-    val differenceFromPrevious: Double?
+    val differenceFromPrevious: Float?
         get() = previousMostRecentLog?.weight?.value?.let {
             mostRecentLog?.weight?.value
                 ?.minus(it)
         }
 
-    val differenceFromGoal: Double?
+    val differenceFromGoal: Float?
         get() = mostRecentLog?.weight?.value?.let {
             goalWeight.minus(it)
         }
