+++ app/src/main/java/com/leeweeder/weighttracker/util/Weight.kt
@@ -3,10 +3,10 @@ package com.leeweeder.weighttracker.util
 import com.leeweeder.weighttracker.ui.util.formatToOneDecimalPlace
 
 data class Weight(
-    val value: Double
+    val value: Float
 ) {
     val displayValue: String
         get() = value.formatToOneDecimalPlace()
 }
 
-const val MAX_WEIGHT = 450.0
+const val MAX_WEIGHT = 450f
